\name{kmeansstep-package}
\alias{kmeansstep-package}
\alias{kmeansstep}
\docType{package}
\title{
\packageTitle{kmeansstep}
}
\description{
\packageDescription{kmeansstep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmeansstep}
\packageIndices{kmeansstep}
}
\author{
\packageAuthor{kmeansstep}

Maintainer: \packageMaintainer{kmeansstep}
}
\references{
http://sherrytowers.com/2013/10/24/k-means-clustering/
}
\keyword{ package }
\seealso{
}
\examples{

}
