\name{kmeansstep-package}
\alias{kmeansstep-package}
\alias{kmeansstep}
\docType{package}
\title{
\packageTitle{kmeansstep}
}
\description{
\packageDescription{kmeansstep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmeansstep}
\packageIndices{kmeansstep}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{kmeansstep}

Maintainer: \packageMaintainer{kmeansstep}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
